requirements/base.txt
Django==3.2.12
django-environ==0.8.1
psycopg2==2.9.3
djangorestframework==3.13.1

celery==5.2.3
django-celery-results==2.3.0
django-celery-beat==2.2.1

whitenoise==6.0.0

django-filter==21.1
django-extensions==3.1.5
django-cors-headers==3.10.0

drf-jwt==1.19.2

boto3==1.20.20
